refactor: split lab rig materialization
