+++ src/core/runner/mod.rs
@@ -19,6 +19,7 @@ mod lab;
 mod offload_changed_since;
 mod offload_metadata;
 mod resource_metrics;
+mod rig_materialization;
 mod session;
 mod worker;
 mod workspace;
